requirements.txt
pyrogram
tgcrypto
motor
fastapi
uvicorn
python-dotenv
aiohttp
dnspython
pydantic
shortuuid